:hammer: Migrate rt_au_score to rt_hb_score: restructure modules, remove unused files, and update versioning
