+++ __init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
 try:
-    from .src.rt_au_score import *
+    from .src.rt_hb_score import *
 except ImportError:
-    from src.rt_au_score import *
+    from src.rt_hb_score import *
