+++ examples/simple/main.py
@@ -6,7 +6,7 @@ import json
 import logging
 from pathlib import Path
 
-from rt_au_score.modules.preprocessing import Preprocessor
+from rt_hb_score.modules.preprocessing import Preprocessor
 
 logger = logging.getLogger(__name__)
