+++ pyproject.toml
@@ -3,15 +3,15 @@ requires = ["setuptools>=70.3.0,<80.0.0", "setuptools-scm>=8.0.4,<9.0.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "rt_au_score"
+name = "rt_hb_score"
 authors = [
 	{ name = "Batkhuu Byambajav", email = "batkhuu10@gmail.com" },
 	{ name = "Asadbek Kiyomov", email = "asadbey0331@gmail.com" },
 ]
-description = "'rt_au_score' is a python module that score web challenge."
+description = "'rt_hb_score' is a python module that score web challenge."
 readme = "README.md"
 requires-python = ">=3.9,<4.0"
-keywords = ["rt_au_score", "model", "module"]
+keywords = ["rt_hb_score", "model", "module"]
 license = { file = "LICENSE.txt" }
 classifiers = [
 	"Development Status :: 4 - Beta",
@@ -30,11 +30,11 @@ dynamic = ["version", "dependencies", "optional-dependencies"]
 
 # [tool.setuptools.packages.find]
 # where = ["src"]
-# include = ["rt_au_score*"]
+# include = ["rt_hb_score*"]
 # namespaces = false
 
 [tool.setuptools.dynamic]
-version = { attr = "rt_au_score.__version__.__version__" }
+version = { attr = "rt_hb_score.__version__.__version__" }
 dependencies = { file = "./requirements.txt" }
 
 [project.urls]
