+++ scripts/get-version.sh
@@ -17,7 +17,7 @@ fi
 
 ## --- Variables --- ##
 # Load from envrionment variables:
-VERSION_FILE_PATH="${VERSION_FILE_PATH:-./src/rt_au_score/__version__.py}"
+VERSION_FILE_PATH="${VERSION_FILE_PATH:-./src/rt_hb_score/__version__.py}"
 ## --- Variables --- ##
