+++ src/rt_hb_score/heuristics/mouse_events/_base.py
