+++ src/rt_hb_score/heuristics/mouse_events/checkbox_path/checkbox_sequence.py
