+++ src/rt_hb_score/heuristics/mouse_events/checkbox_path/session_time.py
