+++ src/rt_hb_score/heuristics/mouse_events/compare/config.py
