+++ src/rt_hb_score/heuristics/mouse_events/velocity/config.py
