+++ src/rt_hb_score/preprocessing/feature_engineer/checkboxes/_checkbox_events.py
